16/04/2025
--> Added find alternative product under nutrition search
--> Added algorithm that after the API reduces the filter of products, compare products to find the healthiest alternative
--> Fixed footer nav bars to be in sync with design
--> Added searchability to add food page
--> Added a product filter API function
--> Added calorie display on calorie page - WIP

Next Features:
--> Finish of calories displaying
--> Have option to choose display date
--> Have option for calorie save date
--> Work on statistical graphs pages
--> Add date time picker on calorie page
--> Work gym page

+++ lib/Models/database_model.dart
@@ -9,7 +9,7 @@ class CalorieCount {
   String dayOfWeek; // Field for the day (e.g., Monday, Tuesday, etc.)
   int weeklyTarget; // Field for the weekly target
   int dailyTarget; // Field for the daily target
-  int calorieTotals; // Field for the total calories
+  double calorieTotals; // Field for the total calories
   double remainingCaloriesDaily; // Field for the remaining daily calories
   double remainingCaloriesWeekly; // Field for the remaining weekly calories
   double progressDaily; // Field for the daily progress
